+++ test_tf2/test/test_convert.cpp
@@ -37,7 +37,7 @@
 #include <gtest/gtest.h>
 #include <tf2/convert.h>
 #include <tf2_kdl/tf2_kdl.h>
-#include <tf2_bullet/tf2_bullet.h>
+#include <tf2_bullet/tf2_bullet.hpp>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 
 TEST(tf2Convert, kdlToBullet)
