+++ tf2_bullet/test/test_tf2_bullet.cpp
@@ -29,7 +29,7 @@
 /** \author Wim Meeussen */
 
 
-#include <tf2_bullet/tf2_bullet.h>
+#include <tf2_bullet/tf2_bullet.hpp>
 #include <rclcpp/rclcpp.hpp>
 #include <gtest/gtest.h>
 #include <tf2/convert.h>
